Switch to Multi File State Machine
