+++ src/main.cpp
@@ -8,7 +8,10 @@
  * @copyright Copyright MRDT 2023 - All Rights Reserved
  ******************************************************************************/
 
-#include "../examples/OpenCV/TagGenerator.hpp"
+// #include "../examples/OpenCV/TagGenerator.hpp"
+
+#include "./Autonomy_Globals.h"
+#include "./interfaces/StateMachine.hpp"
 
 /******************************************************************************
  * @brief Autonomy main function.
@@ -21,7 +24,13 @@
 int main()
 {
     // This leaks memory when the generateImageMarker function is called lol. Damn OpenCV, never knew it was so bad.
-    GenerateOpenCVArucoMarker(cv::aruco::DICT_4X4_50, 1);
+    // GenerateOpenCVArucoMarker(cv::aruco::DICT_4X4_50, 1);
+
+    InitializeAutonomyLoggers();
+
+    StateMachine states_machine;
+    states_machine.initiate();
+    states_machine.process_event(Idle_NavigatingTransition());
 
     return 0;
 }
