+++ src/states/ApproachingGateState.hpp
@@ -1,18 +1,23 @@
 /******************************************************************************
  * @brief
  *
- * @file ApproachingGateState.cpp
+ * @file ApproachingGateState.hpp
  * @author Eli Byrd (edbgkk@mst.edu)
- * @date 2023-0716
+ * @date 07-17-2023
  *
  * @copyright Copyright MRDT 2023 - All Rights Reserved
  ******************************************************************************/
 
-#include "./StateMachine.hpp"
-
+/******************************************************************************
+ * @brief Approaching Gate State
+ *
+ *
+ * @author Eli Byrd (edbgkk@mst.edu)
+ * @date 07-17-2023
+ ******************************************************************************/
 struct ApproachingGateState : sc::simple_state<ApproachingGateState, StateMachine>
 {
-        ApproachingGateState() { PLOGI_(AutonomyLogger::AL_ConsoleLogger) << "In State: Approaching Gate\n"; }
+        ApproachingGateState() { PLOGI_(AutonomyLogger::AL_ConsoleLogger) << "In State: Approaching Gate"; }
 
         typedef mpl::list<sc::custom_reaction<ApproachingGate_MarkerLostTransition>,
                           sc::custom_reaction<ApproachingGate_AbortTransition>,
@@ -28,15 +33,15 @@ struct ApproachingGateState : sc::simple_state<ApproachingGateState, StateMachin
 
 struct ApproachingGate_MarkerLostTransition : sc::event<ApproachingGate_MarkerLostTransition>
 {
-        ApproachingGate_MarkerLostTransition() { PLOGI_(AutonomyLogger::AL_ConsoleLogger) << "In Transition: Approaching Gate (Marker Lost)\n"; }
+        ApproachingGate_MarkerLostTransition() { PLOGI_(AutonomyLogger::AL_ConsoleLogger) << "In Transition: Approaching Gate (Marker Lost)"; }
 };
 
 struct ApproachingGate_AbortTransition : sc::event<ApproachingGate_AbortTransition>
 {
-        ApproachingGate_AbortTransition() { PLOGI_(AutonomyLogger::AL_ConsoleLogger) << "In Transition: Approaching Gate (Abort)\n"; }
+        ApproachingGate_AbortTransition() { PLOGI_(AutonomyLogger::AL_ConsoleLogger) << "In Transition: Approaching Gate (Abort)"; }
 };
 
 struct ApproachingGate_ReachedMarkerTransition : sc::event<ApproachingGate_ReachedMarkerTransition>
 {
-        ApproachingGate_ReachedMarkerTransition() { PLOGI_(AutonomyLogger::AL_ConsoleLogger) << "In Transition: Approaching Gate (Reached Marker)\n"; }
+        ApproachingGate_ReachedMarkerTransition() { PLOGI_(AutonomyLogger::AL_ConsoleLogger) << "In Transition: Approaching Gate (Reached Marker)"; }
 };
